% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIncubationMarker.R
\name{runIncubationMarker}
\alias{runIncubationMarker}
\title{Run the IncubationMarker app}
\usage{
runIncubationMarker()
}
\arguments{
\item{none}{there is no parameter}
}
\description{
This function is another way to run the app
}
\examples{
runIncubationMarker()
}
\keyword{shiny}
