% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whereHOBO.R
\name{whereHOBO}
\alias{whereHOBO}
\title{Where should I put my tables from HOBO?}
\usage{
whereHOBO()
}
\description{
This function returns the directory where you should paste your HOBO files
}
\examples{
whereHOBO()
}
\keyword{shiny}
